the pid is failed.

+++ include/calculation/pid.h
@@ -11,13 +11,13 @@ typedef struct {
 	int kp;
 	int ti;
 	int td;
-	unsigned long target;
-	unsigned long value;
-	long perror;
-	long ierror;
-	long derror;
-	long last_error;
-	long prev_error;
+	long long target;
+	long long value;
+	long long perror;
+	long long ierror;
+	long long derror;
+	long long last_error;
+	long long prev_error;
 } PID_t;
 
 extern long pid_calc(PID_t *pid);
